Refactor client tests
