+++ openapi-cli/src/test/java/io/ballerina/generators/client/ComparedGeneratedFileTests.java
@@ -18,18 +18,12 @@
 
 package io.ballerina.generators.client;
 
-import io.ballerina.compiler.syntax.tree.AnnotationNode;
-import io.ballerina.compiler.syntax.tree.FunctionBodyNode;
-import io.ballerina.compiler.syntax.tree.NodeList;
 import io.ballerina.compiler.syntax.tree.SyntaxTree;
 import io.ballerina.generators.BallerinaClientGenerator;
 import io.ballerina.generators.OpenApiException;
 import io.ballerina.openapi.cmd.Filter;
 import io.ballerina.openapi.exception.BallerinaOpenApiException;
 import io.ballerina.tools.diagnostics.Diagnostic;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.Operation;
-import io.swagger.v3.oas.models.PathItem;
 import org.ballerinalang.formatter.core.FormatterException;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
@@ -41,23 +35,15 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
-import static io.ballerina.generators.BallerinaClientGenerator.extractDisplayAnnotation;
-import static io.ballerina.generators.BallerinaClientGenerator.generatePathWithPathParameter;
-import static io.ballerina.generators.BallerinaClientGenerator.getFunctionBodyNode;
-import static io.ballerina.generators.BallerinaClientGenerator.getReturnType;
 import static io.ballerina.generators.common.TestUtils.compareGeneratedSyntaxTreeWithExpectedSyntaxTree;
 import static io.ballerina.generators.common.TestUtils.getDiagnostics;
-import static io.ballerina.generators.common.TestUtils.getOpenAPI;
 
 /**
  * All the tests related to the {@link io.ballerina.generators.BallerinaClientGenerator} util.
  */
-public class BallerinaClientGeneratorTests {
+public class ComparedGeneratedFileTests {
     private static final Path RES_DIR = Paths.get("src/test/resources/generators/client").toAbsolutePath();
     private static final Path clientPath = RES_DIR.resolve("ballerina_project/client.bal");
     private static final Path schemaPath = RES_DIR.resolve("ballerina_project/types.bal");
@@ -87,54 +73,6 @@ public class BallerinaClientGeneratorTests {
         List<Diagnostic> diagnostics = getDiagnostics(definitionPath, syntaxTree);
         Assert.assertTrue(diagnostics.isEmpty());
         compareGeneratedSyntaxTreeWithExpectedSyntaxTree(expectedPath, syntaxTree);
-    }
-
-    @Test(description = "Generate Client for path parameter has parameter name as key word - unit tests for method")
-    public void generatePathWithPathParameterTests() {
-        Assert.assertEquals(generatePathWithPathParameter("/v1/v2"), "/v1/v2");
-        Assert.assertEquals(generatePathWithPathParameter("/v1/{version}/v2/{name}"),
-                 "/v1/${'version}/v2/${name}");
-        Assert.assertEquals(generatePathWithPathParameter("/v1/{version}/v2/{limit}"),
-                 "/v1/${'version}/v2/${'limit}");
-        Assert.assertEquals(generatePathWithPathParameter("/v1/{age}/v2/{name}"), "/v1/${age}/v2/${name}");
-    }
-
-    @Test(description = "Tests for returnType")
-    public void getReturnTypeTests() throws IOException, BallerinaOpenApiException {
-        OpenAPI array = getOpenAPI(RES_DIR.resolve("swagger/return_type/all_return_type_operation.yaml"));
-        Assert.assertEquals(getReturnType(array.getPaths().get("/jsonproducts").getGet()), "json|error");
-        Assert.assertEquals(getReturnType(array.getPaths().get("/stringproducts/record").getGet()),
-                "ProductArr|error");
-        Assert.assertEquals(getReturnType(array.getPaths().get("/xmlproducts").getGet()), "XML|error");
-        Assert.assertEquals(getReturnType(array.getPaths().get("/xmlarrayproducts").getGet()), "XMLArr|error");
-    }
-
-    @Test(description = "Display Annotation tests for parameters")
-    public void extractDisplayAnnotationTests() throws IOException, BallerinaOpenApiException {
-        Path definitionPath = RES_DIR.resolve("swagger/openapi_display_annotation.yaml");
-        OpenAPI display = getOpenAPI(definitionPath);
-        Map<String, Object> param01 =
-                display.getPaths().get("/weather").getGet().getParameters().get(0).getExtensions();
-        Map<String, Object> param02 =
-                display.getPaths().get("/weather").getGet().getParameters().get(1).getExtensions();
-        NodeList<AnnotationNode> annotationNodes01 = extractDisplayAnnotation(param01);
-        NodeList<AnnotationNode> annotationNodes02 = extractDisplayAnnotation(param02);
-        Assert.assertEquals(annotationNodes01.get(0).annotValue().orElseThrow().toString().trim(),
-                "{label:\"City name\"}");
-        Assert.assertTrue(annotationNodes02.isEmpty());
-    }
-
-    @Test(description = "Test for header that comes under the parameter section.")
-    public void getHeaderTests() throws IOException, BallerinaOpenApiException {
-        Path definitionPath = RES_DIR.resolve("diagnostic_files/header_parameter.yaml");
-        OpenAPI display = getOpenAPI(definitionPath);
-        Set<Map.Entry<PathItem.HttpMethod, Operation>> operation =
-                display.getPaths().get("/pets").readOperationsMap().entrySet();
-        Iterator<Map.Entry<PathItem.HttpMethod, Operation>> iterator = operation.iterator();
-        FunctionBodyNode bodyNode = getFunctionBodyNode("/pets", iterator.next());
-        Assert.assertEquals(bodyNode.toString(), "{string path=string`/pets`;map<string|string[]>accHeaders=" +
-                "{'X\\-Request\\-ID:'X\\-Request\\-ID,'X\\-Request\\-Client:'X\\-Request\\-Client};_=" +
-                "check self.clientEp-> get(path, accHeaders, targetType=http:Response);}");
     }
 
     @Test(description = "Test openAPI definition to ballerina client source code generation",
