+++ openapi-cli/src/test/resources/testng.xml
@@ -28,14 +28,14 @@ under the License.
             <class name="io.ballerina.ballerina.OpenApiConverterUtilsTest"/>
             <class name="io.ballerina.openapi.cmd.OpenApiGenServiceCmdTest"/>
             <class name="io.ballerina.openapi.CodeGeneratorTest"/>
-            <class name="io.ballerina.generators.BallerinaServiceGeneratorTest"></class>
+            <class name="io.ballerina.generators.service.BallerinaServiceGeneratorTest"></class>
             <class name="io.ballerina.generators.schema.BallerinaSchemaGeneratorTests"></class>
-            <class name="io.ballerina.generators.client.BallerinaClientGeneratorTests"></class>
+            <class name="io.ballerina.generators.client.ComparedGeneratedFileTests"></class>
             <class name="io.ballerina.generators.auth.HttpAuthTests"></class>
             <class name="io.ballerina.generators.auth.OAuth2Tests"></class>
             <class name="io.ballerina.generators.auth.ApiKeyAuthTests"></class>
-            <class name="io.ballerina.generators.client.BallerinaClientGeneratorNegativeTests"></class>
-            <class name="io.ballerina.generators.GeneratorUtilsTests"></class>
+            <class name="io.ballerina.generators.client.ImportsTests"></class>
+            <class name="io.ballerina.generators.common.GeneratorUtilsTests"></class>
             <class name="io.ballerina.generators.client.AnyOfDataTypeTests"></class>
             <class name="io.ballerina.generators.schema.AnyOfDataTypeTests"></class>
             <class name="io.ballerina.generators.schema.OneOfDataTypeTests"></class>
